model/requirements.txt
Flask
opencv-python
ultralytics
python-dotenv
websocket-client